# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sports-Point API",
    "version": "1.0.0",
    "description": "运动会积分与成绩管理系统对外 API 文档。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/attempts": {
      "get": {
        "tags": [
          "成绩"
        ],
        "summary": "查询尝试记录",
        "description": "查询某个运动员或队伍在指定项目的所有尝试记录（含作废状态）。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Attempt"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "athlete_type",
            "in": "query",
            "required": false,
            "description": "运动员类型：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "athlete_ref_id",
            "in": "query",
            "required": false,
            "description": "运动员数据库 ID",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "description": "队伍 ID",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/v1/attempts/{attempt_id}/void": {
      "put": {
        "tags": [
          "成绩"
        ],
        "summary": "作废/取消作废尝试记录",
        "description": "将指定尝试记录标记为作废或取消作废，并自动重算该对象的最终成绩。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "description": "尝试记录 ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidAttemptRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/VoidAttemptRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "参数错误"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
...
      "Attempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "尝试记录 ID"
          },
          "attempt_number": {
            "type": "integer",
            "description": "第几次录入"
          },
          "rank": {
            "type": "integer",
            "description": "名次"
          },
          "performance": {
            "type": "string",
            "description": "成绩文本"
          },
          "is_void": {
            "type": "integer",
            "description": "是否作废：0=有效 1=作废"
          },
          "created_at": {
            "type": "string",
            "description": "创建时间"
          }
        }
      },
...
      "VoidAttemptRequest": {
        "type": "object",
        "properties": {
          "is_void": {
            "type": "boolean",
            "description": "是否作废：true=作废 false=取消作废"
          }
        }
      }
    }
  }
}